Moved model versions to their own root route
