+++ src/zenml/zen_stores/sql_zen_store.py
@@ -5802,8 +5802,8 @@ class SqlZenStore(BaseZenStore):
 
     def list_model_versions(
         self,
-        model_name_or_id: Union[str, UUID],
         model_version_filter_model: ModelVersionFilterModel,
+        model_name_or_id: Optional[Union[str, UUID]] = None,
     ) -> Page[ModelVersionResponseModel]:
         """Get all model versions by filter.
 
@@ -5816,7 +5816,9 @@ class SqlZenStore(BaseZenStore):
             A page of all model versions.
         """
         with Session(self.engine) as session:
-            model_version_filter_model.set_scope_model(model_name_or_id)
+            if model_name_or_id:
+                model_version_filter_model.set_scope_model(model_name_or_id)
+
             query = select(ModelVersionSchema)
             return self.filter_and_paginate(
                 session=session,
@@ -6058,15 +6060,13 @@ class SqlZenStore(BaseZenStore):
 
     def list_model_version_artifact_links(
         self,
-        model_name_or_id: Union[str, UUID],
-        model_version_name_or_id: Union[str, UUID],
+        model_version_id: Union[UUID],
         model_version_artifact_link_filter_model: ModelVersionArtifactFilterModel,
     ) -> Page[ModelVersionArtifactResponseModel]:
         """Get all model version to artifact links by filter.
 
         Args:
-            model_name_or_id: name or ID of the model containing the model version.
-            model_version_name_or_id: name or ID of the model version containing the link.
+            model_version_id: ID of the model version containing the link.
             model_version_artifact_link_filter_model: All filter parameters including pagination
                 params.
 
@@ -6074,12 +6074,8 @@ class SqlZenStore(BaseZenStore):
             A page of all model version to artifact links.
         """
         with Session(self.engine) as session:
-            # issue: https://github.com/tiangolo/sqlmodel/issues/109
-            model_version_artifact_link_filter_model.set_scope_model(
-                model_name_or_id
-            )
             model_version_artifact_link_filter_model.set_scope_model_version(
-                model_version_name_or_id
+                model_version_id
             )
             if model_version_artifact_link_filter_model.only_artifacts:
                 query = (
@@ -6137,25 +6133,20 @@ class SqlZenStore(BaseZenStore):
 
     def delete_model_version_artifact_link(
         self,
-        model_name_or_id: Union[str, UUID],
-        model_version_name_or_id: Union[str, UUID],
+        model_version_id: UUID,
         model_version_artifact_link_name_or_id: Union[str, UUID],
     ) -> None:
         """Deletes a model version to artifact link.
 
         Args:
-            model_name_or_id: name or ID of the model containing the model version.
-            model_version_name_or_id: name or ID of the model version containing the link.
+            model_version_id: ID of the model version containing the link.
             model_version_artifact_link_name_or_id: name or ID of the model version to artifact link to be deleted.
 
         Raises:
             KeyError: specified ID or name not found.
         """
         with Session(self.engine) as session:
-            self.get_model(model_name_or_id)
-            model_version = self.get_model_version(
-                model_name_or_id, model_version_name_or_id
-            )
+            model_version = self.get_model_version(model_version_id)
             query = select(ModelVersionArtifactSchema).where(
                 ModelVersionArtifactSchema.model_version_id == model_version.id
             )
@@ -6239,15 +6230,13 @@ class SqlZenStore(BaseZenStore):
 
     def list_model_version_pipeline_run_links(
         self,
-        model_name_or_id: Union[str, UUID],
-        model_version_name_or_id: Union[str, UUID],
+        model_version_id: UUID,
         model_version_pipeline_run_link_filter_model: ModelVersionPipelineRunFilterModel,
     ) -> Page[ModelVersionPipelineRunResponseModel]:
         """Get all model version to pipeline run links by filter.
 
         Args:
-            model_name_or_id: name or ID of the model containing the model version.
-            model_version_name_or_id: name or ID of the model version containing the link.
+            model_version_id: name or ID of the model version containing the link.
             model_version_pipeline_run_link_filter_model: All filter parameters including pagination
                 params.
 
@@ -6255,11 +6244,8 @@ class SqlZenStore(BaseZenStore):
             A page of all model version to pipeline run links.
         """
         with Session(self.engine) as session:
-            model_version_pipeline_run_link_filter_model.set_scope_model(
-                model_name_or_id
-            )
             model_version_pipeline_run_link_filter_model.set_scope_model_version(
-                model_version_name_or_id
+                model_version_id
             )
             return self.filter_and_paginate(
                 session=session,
@@ -6270,25 +6256,20 @@ class SqlZenStore(BaseZenStore):
 
     def delete_model_version_pipeline_run_link(
         self,
-        model_name_or_id: Union[str, UUID],
-        model_version_name_or_id: Union[str, UUID],
+        model_version_id: UUID,
         model_version_pipeline_run_link_name_or_id: Union[str, UUID],
     ) -> None:
         """Deletes a model version to pipeline run link.
 
         Args:
-            model_name_or_id: name or ID of the model containing the model version.
-            model_version_name_or_id: name or ID of the model version containing the link.
+            model_version_id: name or ID of the model version containing the link.
             model_version_pipeline_run_link_name_or_id: name or ID of the model version to pipeline run link to be deleted.
 
         Raises:
             KeyError: specified ID not found.
         """
         with Session(self.engine) as session:
-            self.get_model(model_name_or_id)
-            model_version = self.get_model_version(
-                model_name_or_id, model_version_name_or_id
-            )
+            model_version = self.get_model_version(model_version_id)
             query = select(ModelVersionPipelineRunSchema).where(
                 ModelVersionPipelineRunSchema.model_version_id
                 == model_version.id
